Refactor actions into separate files.
