+++ src/reducers/members.js
@@ -1,4 +1,4 @@
-import * as actions from '../actions';
+import * as actions from '../actions/members';
 
 const initialState = {
   people: {},
