+++ src/reducers/photos.js
@@ -1,4 +1,4 @@
-import * as actions from '../actions';
+import * as actions from '../actions/photos';
 
 const initialState = {
   photos: [],
